refactor: Menu의 이름이 바뀌더라도, OrderLineItem에 영향을 주지 않도록 수정
